+++ prisma/migrations/20260723160215_init_repository/migration.sql
@@ -1,0 +1,12 @@
+-- CreateTable
+CREATE TABLE "Repository" (
+    "id" TEXT NOT NULL,
+    "name" TEXT NOT NULL,
+    "status" TEXT NOT NULL DEFAULT 'UPLOADED',
+    "storagePath" TEXT,
+    "sizeBytes" INTEGER NOT NULL,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP(3) NOT NULL,
+
+    CONSTRAINT "Repository_pkey" PRIMARY KEY ("id")
+);