+++ src/app.module.ts
@@ -3,6 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from './config/config.module';
 import { DatabaseModule } from './database/database.module';
+import { RepositoriesModule } from './repositories/repositories.module';
 
 /*A module is a container that groups related parts of your application.
 
@@ -16,7 +17,7 @@ Which services it makes available to other modules
 Without modules, NestJS wouldn't know how your application is organized.*/
 
 @Module({
-  imports: [ConfigModule, DatabaseModule],
+  imports: [ConfigModule, DatabaseModule, RepositoriesModule],
   controllers: [AppController],
   providers: [AppService],
 })