+++ src/repositories/repositories.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { RepositoriesController } from './repositories.controller';
+import { RepositoriesService } from './repositories.service';
+import { DatabaseModule } from '../database/database.module';
+import { StorageModule } from '../storage/storage.module';
+
+@Module({
+  imports: [DatabaseModule, StorageModule],
+  controllers: [RepositoriesController],
+  providers: [RepositoriesService],
+  exports: [RepositoriesService],
+})
+export class RepositoriesModule {}